update README.md and migrate from diskerase/ to workflow/
