--- chapter/18/diskerase/samples/diskerase/submit.log
@@ -1,44 +1,0 @@
-
-3e08ef26-6942-11ec-9f31-6a8086e3250c
-90ddc05a-6942-11ec-aeff-6a8086e3250c
-7bb16838-698f-11ec-98df-6a8086e3250c
-6b761fc6-6990-11ec-b247-6a8086e3250c
-f1a832be-6990-11ec-b664-6a8086e3250c
-8a979e38-6991-11ec-b39f-6a8086e3250c
-cfec8b38-6991-11ec-84e8-6a8086e3250c
-02843f12-6994-11ec-852e-6a8086e3250c
-4a1c0fa8-6994-11ec-bd49-6a8086e3250c
-a36b3840-6994-11ec-8e1a-6a8086e3250c
-dd7fbcae-6994-11ec-aff4-6a8086e3250c
-4ec51b66-6995-11ec-9402-6a8086e3250c
-6483978e-6995-11ec-b882-6a8086e3250c
-8c1bacea-6996-11ec-9d2b-6a8086e3250c
-93f87e56-6997-11ec-a396-6a8086e3250c
-e6edbb1c-6997-11ec-8d1c-6a8086e3250c
-496cd76e-6998-11ec-90a6-6a8086e3250c
-97a70a76-6998-11ec-8d94-6a8086e3250c
-24a7ac9a-699a-11ec-aa71-6a8086e3250c
-bbe8a392-69a0-11ec-aa71-6a8086e3250c
-d1d03624-69a1-11ec-abe2-6a8086e3250c
-ee0a8984-69a1-11ec-abe2-6a8086e3250c
-d9ef2954-69a2-11ec-abe2-6a8086e3250c
-e56ffd94-69a2-11ec-b025-6a8086e3250c
-20bd97d0-69a3-11ec-b025-6a8086e3250c
-5d7f8084-69a3-11ec-b025-6a8086e3250c
-b9692094-69a3-11ec-b025-6a8086e3250c
-23715a46-69a5-11ec-b025-6a8086e3250c
-cb962d0a-69a5-11ec-b025-6a8086e3250c
-dc5ff4b8-69a5-11ec-b025-6a8086e3250c
-e31edb74-69a6-11ec-b025-6a8086e3250c
-26c57a86-69a7-11ec-b16b-6a8086e3250c
-3467c04a-69a7-11ec-b16b-6a8086e3250c
-c014d7b8-69a7-11ec-8632-6a8086e3250c
-d1131610-69a7-11ec-8632-6a8086e3250c
-c92e68a4-69a8-11ec-963b-6a8086e3250c
-d52fcca6-69a8-11ec-963b-6a8086e3250c
-98a0522c-69aa-11ec-963b-6a8086e3250c
-2acadf4e-69ad-11ec-b84e-6a8086e3250c
-07e5d348-69ae-11ec-b84e-6a8086e3250c
-2588037a-69af-11ec-b84e-6a8086e3250c
-466db92c-69af-11ec-b84e-6a8086e3250c
-6b9ab920-69af-11ec-b84e-6a8086e3250c
